tutor_website/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiosqlite==0.19.0
apscheduler==3.10.4
pytz==2023.3.post1
supabase==1.1.1
openai==0.28.1
aiofiles==23.2.1
email-validator==2.1.0